refactor: cardsGenerator 패키지를 card 내부로 이동
