+++ src/main/java/controller/BlackJackController.java
@@ -2,7 +2,7 @@ package controller;
 
 import domain.GameResult;
 import domain.card.Deck;
-import domain.cardsGenerator.RandomCardsGenerator;
+import domain.card.cardsGenerator.RandomCardsGenerator;
 import domain.participant.Dealer;
 import domain.participant.Player;
 import java.util.List;
