+++ src/main/java/domain/card/cardsGenerator/CardsGenerator.java
@@ -1,4 +1,4 @@
-package domain.cardsGenerator;
+package domain.card.cardsGenerator;
 
 import domain.card.Card;
 import java.util.List;
