+++ src/test/java/domain/DeckTest.java
@@ -8,7 +8,7 @@ import domain.card.CardNumber;
 import domain.card.CardShape;
 import domain.card.Cards;
 import domain.card.Deck;
-import domain.cardsGenerator.RandomCardsGenerator;
+import domain.card.cardsGenerator.RandomCardsGenerator;
 import domain.participant.Dealer;
 import domain.participant.Player;
 import java.util.ArrayList;
